// src/app/(dashboard)/profile/types.ts
export interface IProfileFormInput {
  User_Name: string;
  org_name: string;
  org_id: string;
  branch_name: string;
  branch_code: string;
  branch_id: string;
  branch_address: string;
  branch_mobile: string;
  branch_mail: string;
  is_head: string;
  branch_status: string;
  shg_inv: string;
  user_status: string;
  financialYear: string;
}